+++ employee/src/main/java/ru/fors/employee/domain/usecase/AddEmployeeUseCaseImpl.kt
@@ -12,7 +12,7 @@ import ru.fors.employee.EmployeeUser
 import ru.fors.employee.Role
 import ru.fors.employee.api.domain.dto.EmployeeDto
 import ru.fors.employee.api.domain.dto.EmployeeWithRoleDto
-import ru.fors.employee.api.usecase.AddEmployeeUseCase
+import ru.fors.employee.api.domain.AddEmployeeUseCase
 import ru.fors.employee.data.repo.EmployeeRepo
 import ru.fors.employee.data.repo.EmployeeRoleRepo
 import ru.fors.employee.data.repo.EmployeeUserRepo
