+++ employee/src/main/java/ru/fors/employee/domain/usecase/GetEmployeeRoleUseCaseImpl.kt
@@ -2,7 +2,7 @@ package ru.fors.employee.domain.usecase
 
 import org.springframework.stereotype.Component
 import ru.fors.employee.EmployeeRole
-import ru.fors.employee.api.usecase.GetEmployeeRoleUseCase
+import ru.fors.employee.api.domain.GetEmployeeRoleUseCase
 import ru.fors.employee.data.repo.EmployeeRoleRepo
 
 @Component
