+++ employee_api/src/main/java/ru/fors/employee/api/domain/AddEmployeeUseCase.kt
@@ -1,4 +1,4 @@
-package ru.fors.employee.api.usecase
+package ru.fors.employee.api.domain
 
 import ru.fors.employee.Employee
 import ru.fors.employee.api.domain.dto.EmployeeWithRoleDto
