+++ employee_api/src/main/java/ru/fors/employee/api/domain/GetEmployeeRoleUseCase.kt
@@ -1,4 +1,4 @@
-package ru.fors.employee.api.usecase
+package ru.fors.employee.api.domain
 
 import ru.fors.employee.EmployeeRole
