+++ employee_api/src/main/java/ru/fors/employee/api/domain/dto/EmployeeWithRoleDto.kt
